Refactoring

+++ far/codepage_selection.cpp
@@ -601,7 +601,7 @@ intptr_t codepages::EditDialogProc(Dialog* Dlg, intptr_t Msg, intptr_t Param1, v
 				ConfigProvider().GeneralCfg()->SetValue(NamesOfCodePagesKey, strCodePage, strCodePageName);
 
 			// Получаем информацию о кодовой странице
-			if (const auto Info = GetCodePageInfo(static_cast<UINT>(CodePage)))
+			if (const auto Info = GetCodePageInfo(CodePage))
 			{
 				auto Copy = *Info;
 				const auto IsCodePageNameCustom = GetCodePageCustomName(CodePage, Copy.Name);
@@ -777,13 +777,13 @@ bool codepages::IsCodePageSupported(uintptr_t CodePage, size_t MaxCharSize)
 	if (CodePage == CP_DEFAULT || IsStandardCodePage(CodePage))
 		return true;
 
-	const auto Info = GetCodePageInfo(static_cast<UINT>(CodePage));
+	const auto Info = GetCodePageInfo(CodePage);
 	return Info && Info->MaxCharSize <= MaxCharSize;
 }
 
 std::optional<cp_info> codepages::GetInfo(uintptr_t CodePage)
 {
-	const auto Info = GetCodePageInfo(static_cast<UINT>(CodePage));
+	const auto Info = GetCodePageInfo(CodePage);
 	if (!Info)
 		return {};
